+++ frontend/src/metabase/databases/components/DatabaseHelpCard/DatabaseHelpCard.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { jt, t } from "ttag";
 import Settings from "metabase/lib/settings";
-
+import HelpCard from "metabase/components/HelpCard";
 import ExternalLink from "metabase/core/components/ExternalLink";
-import HelpCard from "../HelpCard";
 
 export interface DatabaseHelpCardProps {
   className?: string;
