+++ frontend/src/metabase/databases/components/DatabaseHelpCard/index.ts
